Refactor scrap
