+++ src/scrap/scrap.h
