+++ src/scrap/scrap.pyi
