+++ src/scrap/scrap_sdl2.c
