+++ src/scrap/scrap_win.c
