+++ test/meson.build
@@ -16,8 +16,6 @@ test_files = files(
     'midi_test.py',
     'mixer_music_test.py',
     'mixer_test.py',
-    'scrap_tags.py',
-    'scrap_test.py',
     'sndarray_tags.py',
     'sndarray_test.py',
     'sprite_test.py',
